filter: Fix one bug in filter
The no value of a filter can only be cur_left_of_filter + 4, which means
that my trick of setting no to 0xFFFF does not work. Keep track of how
much is left of filter, and then for the values where we want to stop
the loop no is cur_left_of_filter + 4.

My code currently works if I specify one port, but not with two or more.