filter: Fix filter creation
The old code for creating filters had two errors:

* We set no to 0xFFFF when we wanted to escape the filter. This is not
allowed, maximum value of no can be (bytes left in filter + 4).
* Yes is only allowed to jump past the command (+ arguments). When we
matched on a port, we would jump much further. This caused validation to
fail.

Filter generation is improved. The maximum value (set on the last
dport/sport) of no is now length + 4. In order to be able to jump
further, we insert a JMP command after each port comparison. This
enables us to jump to the end of the filter as soon as we have a match.

Calculating the maximum is no longer straight forward (and different
tricks can be used to extend the filter length), so I reduced the
maximum number of ports to 256 and plan to come back to setting a sane
maximum.