tcp_closer: Add support for close (kill) via proc
Several machines are running kernels that do not support SOCK_DESTROY.
For example, the kernel might be too old, or CONFIG_INET_DIAG_DESTROY is
disabled.

This commit introduces an alternative way to close a TCP-connection.
Instead of using SOCK_DESTROY, we look through /proc for the inode
belonging to the connection and kill the owning process. This approach
does not work well with all applications, but fits together with for
example ssh (each connection spawns a sub-process).