tcp_closer: Add support for idle timer
Add support to tcp_closer for specifying a limit on how long a
connection can go without having received data, before we kill it. We
check tcpi_last_data_recv, the reason for ignoring ACK is that the
ACK-timer can be updated for example by a proxy.

If limit is not set, all matching connections are destroyed.