netlink: Write message when destroy fails
If the destroy operation for some reason fails, we now output the error.
Output will be improved down the line, for example by writing which
connection we could not erase. In order to add this output, we have
added the ACK-flag to the destroy request, EPOLLIN on the socket and a
receive function.